# [SQL][minor] use catalyst type converter in ScalaUdf