@JoshRosen 2015-05-17
LGTM; seems like a pretty straightforward optimization.  I'm a bit new to Spark SQL, though, so I'll wait for committer with SQL experience to do the final sign-off on this.