@yhuai 2015-05-17
LGTM