@yhuai 2015-05-17
I am merging it to master and branch 1.4.